Structure report with table module
This commit changes the way to format data in order to improve readability and stability. Before this, Slack Report uses tab and newline characters to make a table. However, when the length of values for each cell are different, the user should attach white spaces after or before the values. This commit solves this problem by using `table` module, which takes two-dimensional array as input and returns a clean table.

closes #4