Apply a new report structure (#6)
* Fix the broken tests

* Rename function from valueString to formatString

* Merge tests for util.js

* Structure report with table module

This commit changes the way to format data in order to improve readability and stability. Before this, Slack Report uses tab and newline characters to make a table. However, when the length of values for each cell are different, the user should attach white spaces after or before the values. This commit solves this problem by using `table` module, which takes two-dimensional array as input and returns a clean table.

closes #4

* Change attachments for a new report structure

* Change tests for a new report structure

* Update codes to support a new report structure

* Add an usage example

* Update ReadMe

* Bump up to v0.3.0

* Add git repository to package.json